feat: separate out today vs time entries
