+++ src/Form/TimeEntryListFilterFormType.php
@@ -42,10 +42,7 @@ class TimeEntryListFilterFormType extends AbstractType
 
     public function configureOptions(OptionsResolver $resolver)
     {
-        $resolver->setDefaults([
-                                   'data_class' => TimeEntryListFilterModel::class,
-                               ]);
-
+        $resolver->setDefaults(['data_class' => TimeEntryListFilterModel::class]);
         $resolver->setRequired(['timezone']);
     }
 }
