+++ templates/tag/view.html.twig
@@ -33,7 +33,6 @@
     </div>
     {{ form_end(form) }}
 
-
 </div>
 
 <!-- Delete modal -->
