+++ templates/time_entry/partials/_time-entry-small.html.twig
@@ -1,0 +1,43 @@
+{% import 'time/partials.html.twig' as time %}
+{% import 'tag/partials.html.twig' as tags %}
+
+<div
+    class="card-list-item card-small-spacing time-entry js-time-entry d-flex justify-content-between align-items-center"
+    data-id="{{ timeEntry.idString }}"
+>
+    <div class="d-inline-flex flex-column">
+        <div class="task js-task">
+            {% if not timeEntry.over %}
+                <div class="spinner spinner-border spinner-border-sm text-primary js-time-entry-activity js-loading mr-1" role="status">
+                    <span class="sr-only">Loading...</span>
+                </div>
+            {% endif %}
+            {% if timeEntry.assignedToTask %}
+                <a class="js-task-content" data-task-name="{{ timeEntry.task.name }}" data-task-id="{{ timeEntry.task.idString }}" href="{{ path('task_view', {id: timeEntry.task.id}) }}">{{ timeEntry.task.name }}</a>
+            {% else %}
+                <div data-task-name="" data-task-id="" class="js-task-content d-inline-block">No Task</div>
+            {% endif %}
+        </div>
+        <div class="d-flex js-timestamps mt-2">
+            <div class="datetime js-started-at" data-timestamp="{{ timeEntry.startedAt.timestamp * 1000 }}">{{ time.user_datetime(timeEntry.startedAt, app.user) }}</div>
+            <div
+                class="duration datetime js-duration ml-3"
+                data-start="{{ timeEntry.startedAt.timestamp * 1000 }}"
+                data-duration-format="{{ app.user.durationFormat }}"
+                data-active="{{ timeEntry.running }}">
+                {{ time.user_duration(timeEntry.duration, app.user) }}
+            </div>
+        </div>
+        <div class="tag-list js-tag-list many-rows mt-1" data-initial-tags-name="{{ timeEntry.tagNames() }}">
+            <div class="js-tag-list-view">
+                {% for tag in timeEntry.tags %}
+                    {{ tags.view(tag.name, tag.color) }}
+                {% endfor %}
+            </div>
+        </div>
+    </div>
+
+    <div class="">
+        <a href="{{ path('time_entry_view', {id: timeEntry.idString}) }}" class="btn btn-primary js-view">View</a>
+    </div>
+</div>
